daedalus-13 Build pipeline to intergrate vivarium backbone.
- Update requirements.txt

- Update to the pipeline with the new rate table caching system.

- Removed some old NPP files from cache.

- Test population data added.

- Separated population synthesis code into population_synthesis package and deactivated for now.
