+++ config/default_config.yaml
@@ -1,6 +1,5 @@
 paths:
-    # file should have columns -> PID,location,sex,age,ethnicity
-    path_to_pop_file: './data/1000rows_ssm_E08000032_MSOA11_ppp_2011.csv'
+    path_to_pop_file: './data/ssm_E08000032_MSOA11_ppp_2011.csv'
     path_to_mortality_file: './persistent_data/Mortality2011_LEEDS1_2.csv'
     path_to_fertility_file: './persistent_data/Fertility2011_LEEDS1_2.csv'
     path_to_emigration_file: './persistent_data/Emig_2011_2012_LEEDS2.csv'
