+++ requirements.txt
@@ -1,1 +1,11 @@
-vivarium
+vivarium~=0.9.3
+pyyaml~=5.3.1
+pytest~=6.0.1
+daedalus~=0.0.1
+pandas~=0.24.2
+numpy~=1.19.1
+humanleague~=2.1.1
+ukcensusapi~=1.1.6
+microsimulation~=0.0.1
+ukpopulation~=1.2.1
+setuptools~=49.2.0
