--- tests/cache/NM_2006_1_metadata.json
@@ -1,184 +1,0 @@
-{
-  "nomis_table": "NM_2006_1",
-  "description": "Population projections - local authority based by single year of age",
-  "fields": {
-    "GEOGRAPHY": {
-      "2092957699": "England"
-    },
-    "PROJECTED_YEAR": {
-      "2016": 2016,
-      "2017": 2017,
-      "2018": 2018,
-      "2019": 2019,
-      "2020": 2020,
-      "2021": 2021,
-      "2022": 2022,
-      "2023": 2023,
-      "2024": 2024,
-      "2025": 2025,
-      "2026": 2026,
-      "2027": 2027,
-      "2028": 2028,
-      "2029": 2029,
-      "2030": 2030,
-      "2031": 2031,
-      "2032": 2032,
-      "2033": 2033,
-      "2034": 2034,
-      "2035": 2035,
-      "2036": 2036,
-      "2037": 2037,
-      "2038": 2038,
-      "2039": 2039,
-      "2040": 2040,
-      "2041": 2041
-    },
-    "GENDER": {
-      "0": "Total",
-      "1": "Male",
-      "2": "Female"
-    },
-    "C_AGE": {
-      "200": "All Ages",
-      "201": "Aged 0 to 15",
-      "202": "Aged 16+",
-      "203": "Aged 16 to 64",
-      "250": "Aged 16 to 24",
-      "204": "Aged 16 to 17",
-      "205": "Aged 18 to 24",
-      "206": "Aged 18 to 21",
-      "207": "Aged 25 to 49",
-      "208": "Aged 50 to 64",
-      "209": "Aged 65+",
-      "0": "All Ages",
-      "1": "Age 0 - 4",
-      "101": "Age 0",
-      "102": "Age 1",
-      "103": "Age 2",
-      "104": "Age 3",
-      "105": "Age 4",
-      "3": "Aged 5-9",
-      "106": "Age 5",
-      "107": "Age 6",
-      "108": "Age 7",
-      "109": "Age 8",
-      "110": "Age 9",
-      "4": "Aged 10-14",
-      "111": "Age 10",
-      "112": "Age 11",
-      "113": "Age 12",
-      "114": "Age 13",
-      "115": "Age 14",
-      "5": "Aged 15-19",
-      "116": "Age 15",
-      "117": "Age 16",
-      "118": "Age 17",
-      "119": "Age 18",
-      "120": "Age 19",
-      "6": "Aged 20-24",
-      "121": "Age 20",
-      "122": "Age 21",
-      "123": "Age 22",
-      "124": "Age 23",
-      "125": "Age 24",
-      "7": "Aged 25-29",
-      "126": "Age 25",
-      "127": "Age 26",
-      "128": "Age 27",
-      "129": "Age 28",
-      "130": "Age 29",
-      "8": "Aged 30-34",
-      "131": "Age 30",
-      "132": "Age 31",
-      "133": "Age 32",
-      "134": "Age 33",
-      "135": "Age 34",
-      "9": "Aged 35-39",
-      "136": "Age 35",
-      "137": "Age 36",
-      "138": "Age 37",
-      "139": "Age 38",
-      "140": "Age 39",
-      "10": "Aged 40-44",
-      "141": "Age 40",
-      "142": "Age 41",
-      "143": "Age 42",
-      "144": "Age 43",
-      "145": "Age 44",
-      "11": "Aged 45-49",
-      "146": "Age 45",
-      "147": "Age 46",
-      "148": "Age 47",
-      "149": "Age 48",
-      "150": "Age 49",
-      "12": "Aged 50-54",
-      "151": "Age 50",
-      "152": "Age 51",
-      "153": "Age 52",
-      "154": "Age 53",
-      "155": "Age 54",
-      "13": "Aged 55-59",
-      "156": "Age 55",
-      "157": "Age 56",
-      "158": "Age 57",
-      "159": "Age 58",
-      "160": "Age 59",
-      "14": "Aged 60-64",
-      "161": "Age 60",
-      "162": "Age 61",
-      "163": "Age 62",
-      "164": "Age 63",
-      "165": "Age 64",
-      "15": "Aged 65-69",
-      "166": "Age 65",
-      "167": "Age 66",
-      "168": "Age 67",
-      "169": "Age 68",
-      "170": "Age 69",
-      "16": "Aged 70-74",
-      "171": "Age 70",
-      "172": "Age 71",
-      "173": "Age 72",
-      "174": "Age 73",
-      "175": "Age 74",
-      "17": "Aged 75-79",
-      "176": "Age 75",
-      "177": "Age 76",
-      "178": "Age 77",
-      "179": "Age 78",
-      "180": "Age 79",
-      "18": "Aged 80-84",
-      "181": "Age 80",
-      "182": "Age 81",
-      "183": "Age 82",
-      "184": "Age 83",
-      "185": "Age 84",
-      "210": "Aged 85+",
-      "186": "Age 85",
-      "187": "Age 86",
-      "188": "Age 87",
-      "189": "Age 88",
-      "190": "Age 89",
-      "191": "Aged 90+"
-    },
-    "MEASURES": {
-      "20100": "value",
-      "20301": "percent"
-    },
-    "FREQ": {
-      "A": "Annually"
-    }
-  },
-  "geographies": {
-    "TYPE433": "local authorities: county / unitary (as of April 2019)",
-    "TYPE434": "local authorities: district / unitary (as of April 2019)",
-    "TYPE442": "combined authorities",
-    "TYPE446": "local authorities: county / unitary (as of April 2015)",
-    "TYPE448": "local authorities: district / unitary (as of April 2015)",
-    "TYPE462": "former metropolitan counties",
-    "TYPE463": "local authorities: county / unitary (prior to April 2015)",
-    "TYPE464": "local authorities: district / unitary (prior to April 2015)",
-    "TYPE480": "regions",
-    "TYPE499": "countries"
-  }
-}
